backend/requirements.txt
fastapi
uvicorn[standard]
mediapipe
opencv-python
numpy
pydantic-settings
openai
requests
python-multipart
